Add files via upload
Given the function e^((x^2)/spread), this program asks user to input the spread value, the number of rectangles under the curve, and the range of interval. The program will output the correct area value under the curve depending on the user's input. This program uses two functions that are declared prior to the int main() function.